+++ componets/ui/button.js
@@ -1,0 +1,13 @@
+const Bigbutton = ({name}) => {
+    return (
+        
+            <button className="bg-red-700 text-white font-bold p-5">
+            {name}
+            </button>       
+    );
+};
+
+export default Bigbutton;
+
+
+
