+++ lib/media.js
@@ -1,6 +1,10 @@
 import { getStrapiURL } from "./api";
 
 export function getStrapiMedia(media) {
+    if(media == '') {
+
+        return false;
+    }
     const imageUrl = media.url.startsWith("/")
         ? getStrapiURL(media.url)
         : media.url;
