+++ styles/globals.css
@@ -3,4 +3,28 @@
 @tailwind components;
 @tailwind utilities;
 
+.swiper-pagination-bullet {
+    width: 10px;
+    height: 10px;
+    text-align: right;
+    line-height: 20px;
+    font-size: 12px;
+    color: #000;
+    opacity: 1;
+    background: rgba(0, 0, 0, 0.2);
+  }
+  
+  .swiper-pagination-bullet-active {
+    color: #fff;
+    background: #f5ae33 !important;
+  }
+
+ .swiper-container-horizontal>.swiper-pagination-bullets, .swiper-pagination-custom, .swiper-pagination-fraction {
+    bottom: 0px;
+    left: 0;
+    width: 100%;
+    margin-top:10px;
+    text-align:right;
+}
+
